requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
dotenv==0.9.9
gunicorn==21.2.0
idna==3.4
packaging==23.2
Pillow==10.1.0
psycopg==3.1.12
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
google-generativeai==0.7.2
